+++ packages/lytenyte-core/src/components/cells/cell.tsx
@@ -1,21 +1,18 @@
-import { forwardRef, memo, useCallback, type JSX } from "react";
-import { sizeFromCoord, type LayoutCell } from "@1771technologies/lytenyte-shared";
-import { CellDefault } from "./cell-default.js";
+import { forwardRef, memo, useCallback, useMemo, type JSX } from "react";
+import { type LayoutCell } from "@1771technologies/lytenyte-shared";
 import { useCellStyle } from "./use-cell-style.js";
-import { CellSpacerEnd } from "./cell-spacers/cell-spacer-end.js";
-import { CellSpacerStart } from "./cell-spacers/cell-spacer-start.js";
-import { useBounds, useFocus, useRoot } from "../../root/root-context.js";
-import { $colEndBound, $colStartBound } from "../../selectors.js";
+import { useRoot } from "../../root/root-context.js";
 import { useRowMeta } from "../rows/row/context.js";
 import type { Root } from "../../root/root.js";
 import { useMappedEvents } from "../../hooks/use-mapped-events.js";
 import { useGridId } from "../../root/contexts/grid-id.js";
+import { useColumnSettings } from "../../root/contexts/column-settings/column-settings.js";
+import { useStartBounds } from "../../root/contexts/bounds.js";
+import { useFocusNonReactive } from "../../root/contexts/focus-position.js";
 
 export const Cell = forwardRef<HTMLDivElement, Cell.Props>(function Cell(props, forwarded) {
-  const bounds = useBounds();
-  const end = bounds.useValue($colEndBound);
-  const start = bounds.useValue($colStartBound);
-  const focus = useFocus().get();
+  const [start, end] = useStartBounds();
+  const focus = useFocusNonReactive().get();
 
   const isFocused =
     focus?.kind === "cell" &&
@@ -37,45 +34,41 @@ export const Cell = forwardRef<HTMLDivElement, Cell.Props>(function Cell(props, 
 const CellImpl = memo(
   forwardRef<HTMLDivElement, Cell.Props>(function Cell({ cell, ...props }, forwarded) {
     const id = useGridId();
-    const {
-      rtl,
-      base,
-      xPositions,
-      yPositions,
-      api,
-      view,
-      editMode,
-      events,
-      styles,
-      dimensions: { innerWidth },
-    } = useRoot();
-
+    const { xPositions, yPositions, api, view, editMode, events, styles } = useRoot();
+    const settings = useColumnSettings()[cell.id];
     const rowMeta = useRowMeta();
     const row = rowMeta.row;
 
     const column = view.lookup.get(cell.id)! as Root.Column;
+    const Renderer = settings.cellRenderer;
+    const EditRenderer = settings.editRenderer!;
+    const editSetting = settings.editable;
 
-    const Renderer = column.cellRenderer ?? (base as Root.Column).cellRenderer ?? CellDefault;
-    const EditRenderer = column.editRenderer ?? (base as Root.Column).editRenderer!;
+    const { willDisplayEdit, isEditingThis } = useMemo(() => {
+      const isEditing = Boolean(editMode !== "readonly" && rowMeta.isEditing && EditRenderer);
+      const isEditable =
+        row && typeof editSetting === "function"
+          ? editSetting({ api, row, column, rowIndex: cell.rowIndex, colIndex: cell.colIndex })
+          : editSetting;
 
-    const isEditing = Boolean(editMode !== "readonly" && rowMeta.isEditing && EditRenderer);
+      const isEditingThis =
+        isEditable &&
+        ((editMode === "row" && isEditing) || (editMode === "cell" && rowMeta.editColumn === column.id));
 
-    const editSetting = column.editable ?? (base as Root.Column).editable;
-    const isEditable =
-      row && typeof editSetting === "function"
-        ? editSetting({ api, row, column, rowIndex: cell.rowIndex, colIndex: cell.colIndex })
-        : (editSetting ?? false);
-
-    const isEditingThis =
-      isEditable &&
-      ((editMode === "row" && isEditing) || (editMode === "cell" && rowMeta.editColumn === column.id));
-
-    const willDisplayEdit = isEditing && isEditingThis;
-
-    const style = useCellStyle(xPositions, yPositions, cell, rtl, rowMeta.detailHeight, undefined);
-
-    const handlers = useMappedEvents(events.cell, { column, row, api, layout: cell });
-
+      const willDisplayEdit = isEditing && isEditingThis;
+      return { isEditingThis, willDisplayEdit };
+    }, [
+      EditRenderer,
+      api,
+      cell.colIndex,
+      cell.rowIndex,
+      column,
+      editMode,
+      editSetting,
+      row,
+      rowMeta.editColumn,
+      rowMeta.isEditing,
+    ]);
     const changeValue = useCallback(
       (value: unknown) => {
         return rowMeta.changeValue(value, column);
@@ -83,98 +76,89 @@ const CellImpl = memo(
       [column, rowMeta],
     );
 
-    if (!row || cell.isDeadCol) return null;
+    const style = useCellStyle(xPositions, yPositions, cell, rowMeta.detailHeight);
+    const handlers = useMappedEvents(events.cell, { column, row, api, layout: cell });
 
-    if (cell.isDeadRow) return <div style={{ width: sizeFromCoord(cell.colIndex, xPositions) }} />;
+    if (!row || cell.isDeadCol || cell.isDeadRow) return null;
 
     const selected = row.__selected;
     const indeterminate = row.__indeterminate;
 
     const handleFocus = willDisplayEdit && !!EditRenderer;
     return (
-      <>
-        {cell.colFirstEndPin && (
-          <CellSpacerEnd viewportWidth={innerWidth} visibleEndCount={view.endCount} xPositions={xPositions} />
+      <div
+        className={styles?.cell?.className}
+        {...handlers}
+        {...props}
+        ref={forwarded}
+        role="gridcell"
+        aria-colindex={cell.colIndex + 1}
+        tabIndex={isEditingThis ? -1 : 0}
+        onFocus={
+          handleFocus
+            ? (e) => {
+                (props.onFocus ?? handlers.onFocus)?.(e);
+                if (e.isPropagationStopped()) return;
+
+                rowMeta.setActiveEdit((prev) => {
+                  if (!prev) return prev;
+
+                  return { ...prev, column: column.id };
+                });
+              }
+            : (props.onFocus ?? handlers.onFocus)
+        }
+        style={{ ...style, ...(props.style ?? styles?.cell?.style) }}
+        // Data Properties
+        data-ln-type={settings.type}
+        data-ln-rowindex={cell.rowIndex}
+        data-ln-colindex={cell.colIndex}
+        data-ln-colspan={cell.colSpan}
+        data-ln-rowspan={cell.rowSpan}
+        data-ln-colpin={cell.colPin ?? "center"}
+        data-ln-rowpin={cell.rowPin ?? "center"}
+        data-ln-colid={cell.id}
+        data-ln-gridid={id}
+        data-ln-cell
+        data-ln-edit-active={willDisplayEdit || undefined}
+        data-ln-edit-valid={rowMeta.editValidation == null ? undefined : rowMeta.editValidation === true}
+        data-ln-last-top-pin={cell.rowLastPinTop}
+        data-ln-first-bottom-pin={cell.rowFirstPinBottom}
+        data-ln-last-start-pin={cell.colLastStartPin}
+        data-ln-first-end-pin={cell.colFirstEndPin}
+      >
+        {willDisplayEdit && (
+          <EditRenderer
+            api={api}
+            cancel={rowMeta.cancel}
+            commit={rowMeta.commit}
+            changeData={rowMeta.changeData}
+            changeValue={changeValue}
+            rowIndex={cell.rowIndex}
+            colIndex={cell.colIndex}
+            column={column}
+            row={row}
+            layout={cell}
+            editData={rowMeta.editData}
+            editValidation={rowMeta.editValidation!}
+            editValue={api.columnField(column, { ...row, data: rowMeta.editData })}
+          />
         )}
-
-        <div
-          className={styles?.cell?.className}
-          {...handlers}
-          {...props}
-          ref={forwarded}
-          role="gridcell"
-          aria-colindex={cell.colIndex + 1}
-          tabIndex={isEditingThis ? -1 : 0}
-          onFocus={
-            handleFocus
-              ? (e) => {
-                  (props.onFocus ?? handlers.onFocus)?.(e);
-                  if (e.isPropagationStopped()) return;
-
-                  rowMeta.setActiveEdit((prev) => {
-                    if (!prev) return prev;
-
-                    return { ...prev, column: column.id };
-                  });
-                }
-              : (props.onFocus ?? handlers.onFocus)
-          }
-          style={{ ...style, ...(props.style ?? styles?.cell?.style) }}
-          // Data Properties
-          data-ln-type={column.type ?? base.type ?? "string"}
-          data-ln-rowindex={cell.rowIndex}
-          data-ln-colindex={cell.colIndex}
-          data-ln-colspan={cell.colSpan}
-          data-ln-rowspan={cell.rowSpan}
-          data-ln-colpin={cell.colPin ?? "center"}
-          data-ln-rowpin={cell.rowPin ?? "center"}
-          data-ln-colid={cell.id}
-          data-ln-gridid={id}
-          data-ln-cell
-          data-ln-edit-active={willDisplayEdit || undefined}
-          data-ln-edit-valid={rowMeta.editValidation == null ? undefined : rowMeta.editValidation === true}
-          data-ln-last-top-pin={cell.rowLastPinTop}
-          data-ln-first-bottom-pin={cell.rowFirstPinBottom}
-          data-ln-last-start-pin={cell.colLastStartPin}
-          data-ln-first-end-pin={cell.colFirstEndPin}
-        >
-          {willDisplayEdit && (
-            <EditRenderer
-              api={api}
-              cancel={rowMeta.cancel}
-              commit={rowMeta.commit}
-              changeData={rowMeta.changeData}
-              changeValue={changeValue}
-              rowIndex={cell.rowIndex}
-              colIndex={cell.colIndex}
-              column={column}
-              row={row}
-              layout={cell}
-              editData={rowMeta.editData}
-              editValidation={rowMeta.editValidation!}
-              editValue={api.columnField(column, { ...row, data: rowMeta.editData })}
-            />
-          )}
-          {!willDisplayEdit && (
-            <Renderer
-              column={column}
-              editData={rowMeta.editData}
-              row={row}
-              api={api}
-              layout={cell}
-              rowIndex={cell.rowIndex}
-              colIndex={cell.colIndex}
-              selected={selected}
-              indeterminate={indeterminate}
-              detailExpanded={rowMeta.detailExpanded}
-            />
-          )}
-        </div>
-
-        {cell.colLastStartPin && (
-          <CellSpacerStart xPositions={xPositions} rowMeta={rowMeta} visibleStartCount={view.startCount} />
+        {!willDisplayEdit && (
+          <Renderer
+            column={column}
+            editData={rowMeta.editData}
+            row={row}
+            api={api}
+            layout={cell}
+            rowIndex={cell.rowIndex}
+            colIndex={cell.colIndex}
+            selected={selected}
+            indeterminate={indeterminate}
+            detailExpanded={rowMeta.detailExpanded}
+          />
         )}
-      </>
+      </div>
     );
   }),
 );