+++ packages/lytenyte-core/src/components/viewport/view-monitor.tsx
@@ -1,12 +1,15 @@
 import { supportsScrollEnd } from "@1771technologies/lytenyte-shared";
-import { useEffect } from "react";
-import { useBounds, useRoot } from "../../root/root-context.js";
+import { useEffect, useRef } from "react";
+import { useRoot } from "../../root/root-context.js";
+import { useBounds } from "../../root/contexts/bounds.js";
 
 export function ViewMonitor({ viewport }: { viewport: HTMLElement }) {
   const { source } = useRoot();
-  const bounds = useBounds();
+  const boundsValue = useBounds();
 
-  const boundsValue = bounds.useValue();
+  const ref = useRef(boundsValue);
+  ref.current = boundsValue;
+
   useEffect(() => {
     source.onViewChange(boundsValue);
   }, [boundsValue, source]);
@@ -15,13 +18,9 @@ export function ViewMonitor({ viewport }: { viewport: HTMLElement }) {
     const controller = new AbortController();
 
     if (supportsScrollEnd(viewport)) {
-      viewport.addEventListener(
-        "scrollend",
-        () => {
-          source.onViewChange(bounds.get());
-        },
-        { signal: controller.signal },
-      );
+      viewport.addEventListener("scrollend", () => source.onViewChange(ref.current), {
+        signal: controller.signal,
+      });
     } else {
       let timeout: ReturnType<typeof setTimeout> | null = null;
       viewport.addEventListener(
@@ -29,7 +28,7 @@ export function ViewMonitor({ viewport }: { viewport: HTMLElement }) {
         () => {
           if (!timeout)
             timeout = setTimeout(() => {
-              source.onViewChange(bounds.get());
+              source.onViewChange(ref.current);
               timeout = null;
             }, 200);
         },
@@ -38,7 +37,7 @@ export function ViewMonitor({ viewport }: { viewport: HTMLElement }) {
     }
 
     return () => controller.abort();
-  }, [bounds, source, viewport]);
+  }, [source, viewport]);
 
   return null;
 }