+++ packages/lytenyte-core/src/root/root-context.ts
@@ -3,11 +3,8 @@ import type {
   ColumnView,
   DataRect,
   LayoutHeader,
-  PositionUnion,
   RowNode,
   RowSource,
-  RowView,
-  SpanLayout,
 } from "@1771technologies/lytenyte-shared";
 import {
   createContext,
@@ -17,7 +14,7 @@ import {
   type RefObject,
   type SetStateAction,
 } from "react";
-import type { Piece, PieceWritable } from "../hooks/use-piece";
+import type { Piece } from "../hooks/use-piece";
 import type { Dimension } from "./hooks/use-viewport-dimensions";
 import type { Root } from "./root";
 import type { ViewportShadowsProps } from "../components/viewport/viewport-shadows";
@@ -58,8 +55,6 @@ export interface RootContextValue {
 
   readonly styles: Props["styles"];
 
-  readonly focusActive: PieceWritable<PositionUnion | null>;
-
   readonly events: GridEvents<GridSpec>;
 
   readonly columnGroupDefaultExpansion: boolean;
@@ -94,17 +89,9 @@ const context = createContext<RootContextValue>({} as any);
 export const RootContextProvider = context.Provider;
 export const useRoot = () => useContext(context);
 
-const rowContextLayout = createContext<RowView>({} as any);
-export const RowLayoutContextProvider = rowContextLayout.Provider;
-export const useRowLayout = () => useContext(rowContextLayout);
-
 const colContextLayout = createContext<LayoutHeader[][]>({} as any);
 export const ColumnLayoutContextProvider = colContextLayout.Provider;
 export const useColumnLayout = () => useContext(colContextLayout);
-
-const boundsContext = createContext<Piece<SpanLayout>>({} as any);
-export const BoundsContextProvider = boundsContext.Provider;
-export const useBounds = () => useContext(boundsContext);
 
 export interface EditContext {
   readonly activeEdit: null | { readonly rowId: string; readonly column: string };
@@ -126,7 +113,3 @@ export interface EditContext {
 const editContext = createContext({} as EditContext);
 export const EditProvider = editContext.Provider;
 export const useEdit = () => useContext(editContext);
-
-const focusContext = createContext<Piece<PositionUnion | null>>({} as any);
-export const FocusProvider = focusContext.Provider;
-export const useFocus = () => useContext(focusContext);